+++ CHANGELOG.md
@@ -16,9 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Opacity in entity images is dependent on structure size
 - Save _api_data.json
 - Create _expected_files.txt and check for missing files after run
-- Hide hydrogens and branched entity balls-and-sticks by default, new args --show-hydrogens --show-branched-sticks --allow-lowest-quality
+- Hide hydrogens and branched entity balls-and-sticks by default, new args --show-hydrogens --show-branched-sticks --ensemble-shades --allow-lowest-quality
 - Fixed color assignment for ensembles
-- Models in ensemble shown in diffent lightness variants
 
 
 ## [1.0.0] - 2023-06-20